跨链模块拆分
